requirements.txt
requests>=2.28.0
beautifulsoup4>=4.11.0
google-auth>=2.22.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.100.0
